style(services/authservice): renamed to publicFormAuthService
